started v3 of pc
print job 2 also lost adhesion and failed, but it lasted ~30-40 layers instead of five. currently reprinting with gluestick and increased ambient chamber temp (let heat for longer, added towel), and is currently at later 55 with no warping. This commit added a ender display cover to add to the ender to protect from damages from the liquids eventually bring printed from it, and the time-lapse of the failed pc print #2 (really bad bc i only stopped the print at 3am, but it failed at ~1am, leading to a lot of cleanup before the next print)